Changed weather layout completely (retorfit request to single only)

+++ app/src/main/java/com/fyp/agrifarm/app/SettingsFragment.java
@@ -1,9 +1,11 @@
 package com.fyp.agrifarm.app;
 
+import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
 import android.widget.Toast;
 
+import androidx.annotation.NonNull;
 import androidx.preference.Preference;
 import androidx.preference.PreferenceFragmentCompat;
 
@@ -12,14 +14,15 @@ import com.fyp.agrifarm.utils.FirebaseUtils;
 
 public class SettingsFragment extends PreferenceFragmentCompat {
 
+    private OnPreferencesChangeListener listener;
+
     @Override
     public void onCreatePreferences(Bundle savedInstanceState, String rootKey) {
         setPreferencesFromResource(R.xml.main_preferences, rootKey);
 
         Preference deletePreference = findPreference("delete");
-        deletePreference.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
-            @Override
-            public boolean onPreferenceClick(Preference preference) {
+        if (deletePreference != null) {
+            deletePreference.setOnPreferenceClickListener(preference -> {
                 FirebaseUtils.deleteAccount(o -> {
                             Toast.makeText(getContext(), "Account Deleted!", Toast.LENGTH_SHORT).show();
                             startActivity(new Intent(getContext(), LoginActivity.class));
@@ -28,10 +31,31 @@ public class SettingsFragment extends PreferenceFragmentCompat {
                 );
 
                 return true;
+            });
+        }
+
+        Preference weather = findPreference("weatherUnit");
+        weather.setOnPreferenceChangeListener(new Preference.OnPreferenceChangeListener() {
+            @Override
+            public boolean onPreferenceChange(Preference preference, Object newValue) {
+                listener.onWeatherChanged(newValue.toString());
+                return true;
             }
         });
-
-
     }
+
+    @Override
+    public void onAttach(@NonNull Context context) {
+        super.onAttach(context);
+
+        listener = (OnPreferencesChangeListener) context;
+    }
+
+    public interface OnPreferencesChangeListener {
+        void onWeatherChanged(String newValue);
+    }
+
+
+    
 
 }
